requirements.txt
llama-cpp-python==0.1.77
flask
PyPDF2
pysimilar
langchain

